Внес правки в функции и изменил структуру проекта
