+++ DataBase/Procedures.php
@@ -1,0 +1,10 @@
+<?php
+
+/**Список хранимых процедур из БД MySQL*/
+class Procedures
+{
+    /**Получить типы контента*/
+    public const  sqlTypeContent = "CALL GetTypeContent";
+    /**Получить список постов*/
+    public const  sqlPostUserType = "CALL GetPostUserType";
+}
